Wip animations events

+++ Assets/Scripts/Domain/AnimationStates/Attack.cs
@@ -6,5 +6,5 @@ public class Attack
     [field: SerializeField] public string Animation { get; set; }
     [field: SerializeField] public int Damage { get; set; }
     [field: SerializeField] public int Range { get; set; }
-    [field: SerializeField] public int Delay { get; set; }
+    [field: SerializeField] public float Delay { get; set; }
 }
